backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
flwr==1.7.0
llama-cpp-python==0.2.28
numpy==1.26.3
python-multipart==0.0.6
sqlite3
datetime
typing
